refactor(governance): move machine identity routes
